pyproject.toml
[tool.poetry]
name = "flightscanner"
version = "0.1.0"
description = ""
authors = ["Felipe Carvalho <felipe.carvalho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
deprecated = "^1.2.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
